requirements.txt
asn1crypto==0.24.0
autopep8==1.3.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
defusedxml==0.5.0
Django==2.0.7
django-admin-interface==0.6.3
django-allauth==0.36.0
django-colorfield==0.1.14
django-cors-middleware==1.3.1
django-debug-toolbar==1.9.1
django-extensions==2.1.0
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.11.1
django-gravatar2==1.4.2
django-jet==1.0.7
django-mysql==2.4.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-suit==2.0a1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
jwt==0.5.4
Markdown==2.6.11
mysql-connector-python==8.0.12
oauthlib==2.1.0
pep8==1.7.1
protobuf==3.6.0
psycopg2==2.7.5
pycodestyle==2.4.0
pycparser==2.18
PyJWT==1.6.4
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
var-dump==1.2
